add common utils module inherited by fina_utils and api_utils
